refactor equipment management
